+++ lib/src/config/v2/tui/mod.rs
@@ -1,4 +1,4 @@
-use std::path::Path;
+use std::{collections::HashSet, path::Path};
 
 use anyhow::{Context, Result};
 use serde::{Deserialize, Serialize};
@@ -100,6 +100,53 @@ pub struct CoverArt {
     pub size_scale: i8,
     /// Whether to show or hide the coverart if it is compiled in
     pub hidden: bool,
+
+    /// Enabled coverart display protocols. Protocols not compiled-in will not have a effect.
+    ///
+    /// Remove items from this list to disable the protocol
+    pub protocols: CoverArtProtocolsSet,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
+#[serde(transparent)]
+pub struct CoverArtProtocolsSet(HashSet<CoverArtProtocol>);
+
+impl Default for CoverArtProtocolsSet {
+    fn default() -> Self {
+        Self(HashSet::from(*PROTOCOLS_DEFAULT))
+    }
+}
+
+impl CoverArtProtocolsSet {
+    /// Check if a given [`CoverArtProtocol`] is enabled.
+    #[inline]
+    #[must_use]
+    pub fn includes_protocol(&self, protocol: CoverArtProtocol) -> bool {
+        self.0.contains(&protocol)
+    }
+}
+
+/// All protocols are enabled by default.
+pub const PROTOCOLS_DEFAULT: &[CoverArtProtocol; 4] = &[
+    CoverArtProtocol::Kitty,
+    CoverArtProtocol::Iterm2,
+    CoverArtProtocol::Sixel,
+    CoverArtProtocol::Ueberzug,
+];
+
+/// All available Cover-Art protocols in the TUI.
+///
+/// Only has a effect if said protocol is compiled-in.
+#[derive(Debug, Clone, Copy, Deserialize, Serialize, PartialEq, Eq, Hash)]
+pub enum CoverArtProtocol {
+    #[serde(rename = "sixel")]
+    Sixel,
+    #[serde(rename = "iterm2", alias = "iterm")]
+    Iterm2,
+    #[serde(rename = "kitty")]
+    Kitty,
+    #[serde(rename = "ueberzug")]
+    Ueberzug,
 }
 
 #[derive(Debug, Clone, Copy, Deserialize, Serialize, Default, PartialEq, Eq)]
@@ -124,7 +171,7 @@ pub struct Ytdlp {
 
 mod v1_interop {
     use super::{Alignment, BehaviorSettings, CoverArt, MaybeComSettings, TuiSettings, Ytdlp};
-    use crate::config::v1;
+    use crate::config::{v1, v2::tui::CoverArtProtocolsSet};
 
     impl From<v1::Alignment> for Alignment {
         fn from(value: v1::Alignment) -> Self {
@@ -145,6 +192,7 @@ mod v1_interop {
                 // the value is named "width", but more use like a scale on both axis
                 size_scale: value.width_between_1_100.clamp(0, i8::MAX as u32) as i8,
                 hidden: Self::default().hidden,
+                protocols: CoverArtProtocolsSet::default(),
             }
         }
     }
@@ -190,7 +238,8 @@ mod v1_interop {
                 CoverArt {
                     align: Alignment::BottomRight,
                     size_scale: 20,
-                    hidden: false
+                    hidden: false,
+                    protocols: CoverArtProtocolsSet::default()
                 }
             );